+++ src/gql/types/generators/AssetType.php
@@ -8,10 +8,9 @@
 namespace craft\gql\types\generators;
 
 use Craft;
-use craft\base\Field;
-use craft\base\GqlSchemaAwareFieldInterface;
 use craft\base\Volume;
 use craft\elements\Asset as AssetElement;
+use craft\gql\base\BaseGenerator;
 use craft\gql\base\GeneratorInterface;
 use craft\gql\base\ObjectType;
 use craft\gql\base\SingleGeneratorInterface;
@@ -28,7 +27,7 @@ use craft\helpers\Gql as GqlHelper;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 3.3.0
  */
-class AssetType implements GeneratorInterface, SingleGeneratorInterface
+class AssetType extends BaseGenerator implements GeneratorInterface, SingleGeneratorInterface
 {
     /**
      * @inheritdoc
@@ -60,15 +59,7 @@ class AssetType implements GeneratorInterface, SingleGeneratorInterface
     {
         /** @var Volume $volume */
         $typeName = AssetElement::gqlTypeNameByContext($context);
-        $contentFields = $context->getFields();
-        $contentFieldGqlTypes = [];
-
-        /** @var Field $contentField */
-        foreach ($contentFields as $contentField) {
-            if (!$contentField instanceof GqlSchemaAwareFieldInterface || $contentField->getExistsForCurrentGqlSchema()) {
-                $contentFieldGqlTypes[$contentField->handle] = $contentField->getContentGqlType();
-            }
-        }
+        $contentFieldGqlTypes = self::getContentFields($context);
 
         $assetFields = TypeManager::prepareFieldDefinitions(array_merge(AssetInterface::getFieldDefinitions(), $contentFieldGqlTypes), $typeName);