+++ src/gql/types/generators/MatrixBlockType.php
@@ -12,6 +12,7 @@ use craft\base\Field;
 use craft\base\GqlSchemaAwareFieldInterface;
 use craft\elements\MatrixBlock as MatrixBlockElement;
 use craft\fields\Matrix;
+use craft\gql\base\BaseGenerator;
 use craft\gql\base\GeneratorInterface;
 use craft\gql\base\ObjectType;
 use craft\gql\base\SingleGeneratorInterface;
@@ -27,7 +28,7 @@ use craft\models\MatrixBlockType as MatrixBlockTypeModel;
  * @author Pixel & Tonic, Inc. <support@pixelandtonic.com>
  * @since 3.3.0
  */
-class MatrixBlockType implements GeneratorInterface, SingleGeneratorInterface
+class MatrixBlockType extends BaseGenerator implements GeneratorInterface, SingleGeneratorInterface
 {
     /**
      * @inheritdoc
@@ -62,16 +63,7 @@ class MatrixBlockType implements GeneratorInterface, SingleGeneratorInterface
         $typeName = MatrixBlockElement::gqlTypeNameByContext($context);
 
         if (!($entity = GqlEntityRegistry::getEntity($typeName))) {
-            $contentFields = $context->getFields();
-            $contentFieldGqlTypes = [];
-
-            /** @var Field $contentField */
-            foreach ($contentFields as $contentField) {
-                if (!$contentField instanceof GqlSchemaAwareFieldInterface || $contentField->getExistsForCurrentGqlSchema()) {
-                    $contentFieldGqlTypes[$contentField->handle] = $contentField->getContentGqlType();
-                }
-            }
-
+            $contentFieldGqlTypes = self::getContentFields($context);
             $blockTypeFields = TypeManager::prepareFieldDefinitions(array_merge(MatrixBlockInterface::getFieldDefinitions(), $contentFieldGqlTypes), $typeName);
 
             // Generate a type for each block type